Add Web Developer role and update programming languages
This commit includes two main changes to the project. Firstly, the
commented-out line for the Web Developer role in index.html has been
uncommented, officially adding the Web Developer role to the list of
professional skills. Secondly, the list of programming languages has
been updated to include Java at the beginning, reflecting a broader
range of expertise. Additionally, a new image file, DSC_6267 copy.png,
has been added to the images directory, although the specific purpose
of this file is not detailed in this commit.