Add NVIDIA certificate and update skills section
This commit introduces several updates to the portfolio section. Firstly, it adds a new NVIDIA certificate to the certificates section, including an image and a link to the certificate. The image `Nvidia.png` is added to the project to support this update.

Additionally, the skills section in `index.html` has been restructured for better clarity and presentation. The "Developer Skills" and "Soft Skills" are now presented in separate cards, enhancing the visual layout and making the information more accessible. A minor punctuation inconsistency in the "Web development" skills list has also been corrected by removing an extraneous period.

Lastly, a commented-out section related to NVIDIA experience has been removed to clean up the code and focus on the current content.

These changes aim to improve the portfolio's content organization and showcase new credentials effectively.