Update project details and reorganize portfolio
This commit introduces several changes to the portfolio section of the website. Firstly, it comments out a project card that might be under revision or no longer showcased. Additionally, it updates project categories and titles to reflect current or new projects, such as changing a project from "Portfolio" to "Reddit Clone" and updating its associated technologies and description. This change aims to keep the portfolio fresh and relevant.

Moreover, a new project titled "Vector Database Indexer" is added, showcasing skills in Python, algorithms, and data structures. This addition enriches the portfolio with a project that emphasizes backend capabilities and complex problem-solving skills.

The commit also reintroduces a previously listed project, "E-Commerce System," at a different position in the document, possibly for better visibility or organizational purposes. Lastly, it updates another project from "Vector Database Indexer" to "Logic Simulator," reflecting a shift in showcased skills from Python and algorithms to C++ and object-oriented programming concepts.

These changes collectively aim to present a more accurate and updated representation of skills and project experiences on the website.