Update GPA, skills, and styling adjustments
This commit includes several updates to the portfolio content and styling. The GPA has been corrected from 3.6 to 3.7 to reflect the most recent academic achievement. In the skills section, R has been removed from the list of programming languages, and the skills categories have been reorganized to better represent current capabilities, including the addition of a Back-end category with Java, a Databases category, and adjustments to the AI and Data Analysis sections. The Mobile and Web development sections have been updated to reflect current focus areas.

Styling adjustments have been made across the CSS file to improve the responsiveness and visual appeal of the portfolio. These changes include adjustments to margins, font sizes, and background images for better readability and layout on various screen sizes. The aim is to provide a more engaging and accessible user experience.