Update back-end technologies and adjust CSS
This commit adds PHP and Laravel to the list of back-end technologies
mentioned in the index.html file, under the Back-end section. It also
removes unnecessary styling for the .programming-lang class in the
style.css file, streamlining the CSS codebase. These changes aim to
reflect our current back-end capabilities more accurately and improve
the website's maintainability.